Session 1
Stopped at lecture 20. Mostly setup, learning how to use best
practices, and find functions in the WP codex. Will cover how to
properly add CDN content in a later lesson. Added lots of comments to
index.php and will likely continue that in other files to serve as a
reference.